fix: check if URL with replaced defaultLocale in notFoundRoutes
We currently replace the defaultLocale (when i18n is enabled) in the
urlSet (with the help of a regex), but URLs in notFoundRoutes are not
replaced.

This would cause notFound routes on the defaultLocale to be missed when
filtering.

Add additional check to check if url with edfaultLocale replacedd is
included in notFoundRoutes